Add per core visualization

+++ Visualizer/Source/Visualizer.py
@@ -13,6 +13,14 @@ from Visualizer.Testing.TestResults import TestResults
 def parse_arguments():
     parser = argparse.ArgumentParser(prog="Visualizer", description="Visualize EnergyManager test results.")
 
+    parser.add_argument(
+        "--update",
+        "-u",
+        action="store_true",
+        help="Updates existing results (may take a long time).",
+        required=False
+    )
+
     input_group = parser.add_argument_group("input")
     input_group.add_argument(
         "--database",
@@ -42,20 +50,55 @@ if __name__ == '__main__':
     arguments = parse_arguments()
 
     for test in Test.load_all(arguments.database):
-        # Determine the output directory for the current test
-        output_directory = f"{arguments.output_directory}/{test.name}"
+        print(f"Processing test {test.id} - {test.name}...")
 
-        # Make sure the directory exists and is empty
+        # Determine the output directory for the current test
+        output_directory = f"{arguments.output_directory}/{test.id} - {test.name}"
+
+        # Check if the results already exists
         if os.path.exists(output_directory):
-            shutil.rmtree(output_directory)
+            # If so, check if we need to update them
+            if arguments.update:
+                print(f"Results exist, updating...")
+
+                # Delete the old results
+                shutil.rmtree(output_directory)
+            else:
+                print(f"Results exist, skipping...")
+
+                # Go to the next test
+                continue
+
+        # Create a directory to hold the results
         os.makedirs(output_directory)
+
 
         def collect_values(test_results: TestResults, monitor: str, name: str, type, modifier=lambda value: value):
             values: OrderedDict[int, type] = collections.OrderedDict()
             for timestamp, results in test_results.monitor_results[monitor].items():
-                values[timestamp] = modifier(type(results[name]))
+                try:
+                    values[timestamp] = modifier(type(results[name]))
+                except:
+                    pass
 
             return values
+
+
+        def collect_indexed_values(test_results: TestResults, monitor: str, name: str, type, modifier=lambda value: value):
+            index = 0
+
+            results = list()
+
+            while True:
+                values = collect_values(test_results, monitor, name + str(index), type, modifier)
+
+                if len(values) == 0:
+                    break
+                else:
+                    results.append(values)
+                    index += 1
+
+            return results
 
 
         def plot(test: Test, test_results: TestResults, size: Tuple[int, int]):
@@ -70,13 +113,32 @@ if __name__ == '__main__':
             j_to_wh = lambda value: value / 3600
             to_percentage = lambda value: value * 100
             ns_to_s = lambda value: value / 1e9
+
+            def add_indexed_value(plot, series_name_prefix: str, monitor: str, name: str, type, modifier = lambda value: value):
+                for index, values in enumerate(collect_indexed_values(test_results, monitor, name, type, modifier)):
+                    plot[(f"{series_name_prefix} {index}", "")] = values
+
+            clock_rate_plot = {
+                ("CPU", ""): collect_values(test_results, "CPUMonitor", "coreClockRate", int),
+            }
+            add_indexed_value(clock_rate_plot, "CPU Core", "CPUMonitor", "coreClockRateCore", int)
+            clock_rate_plot.update({
+                ("GPU Core", ""): collect_values(test_results, "GPUMonitor", "coreClockRate", int),
+                ("GPU Memory", ""): collect_values(test_results, "GPUMonitor", "memoryClockRate", int),
+                ("GPU SM", ""): collect_values(test_results, "GPUMonitor", "streamingMultiprocessorClockRate", int),
+            })
+
+            utilization_rate_plot = {
+                ("CPU", ""): collect_values(test_results, "CPUMonitor", "coreUtilizationRate", float, to_percentage),
+            }
+            add_indexed_value(utilization_rate_plot, "CPU Core", "CPUMonitor", "coreUtilizationRateCore", float, to_percentage)
+            utilization_rate_plot.update({
+                ("GPU Core", ""): collect_values(test_results, "GPUMonitor", "coreUtilizationRate", float, to_percentage),
+                ("GPU Memory", ""): collect_values(test_results, "GPUMonitor", "memoryUtilizationRate", float, to_percentage),
+            })
+
             plots = {
-                "Clock Rate (Hz)": {
-                    ("CPU Core", ""): collect_values(test_results, "CPUMonitor", "coreClockRate", int),
-                    ("GPU Core", ""): collect_values(test_results, "GPUMonitor", "coreClockRate", int),
-                    ("GPU Memory", ""): collect_values(test_results, "GPUMonitor", "memoryClockRate", int),
-                    ("GPU SM", ""): collect_values(test_results, "GPUMonitor", "streamingMultiprocessorClockRate", int),
-                },
+                "Clock Rate (Hz)": clock_rate_plot,
                 "Energy Consumption (J)": {
                     ("CPU", ""): collect_values(test_results, "CPUMonitor", "energyConsumption", float),
                     ("GPU", ""): collect_values(test_results, "GPUMonitor", "energyConsumption", float),
@@ -87,8 +149,8 @@ if __name__ == '__main__':
                     ("GPU", ""): collect_values(test_results, "GPUMonitor", "energyConsumption", float, j_to_wh),
                     ("Node", ""): collect_values(test_results, "NodeMonitor", "energyConsumption", float, j_to_wh),
                 },
-                "Fan Speed (RPM)": {
-                    ("GPU", ""): collect_values(test_results, "GPUMonitor", "fanSpeed", float),
+                "Fan Speed (%)": {
+                    ("GPU", ""): collect_values(test_results, "GPUMonitor", "fanSpeed", float, to_percentage),
                 },
                 "Power Consumption (W)": {
                     ("CPU", ""): collect_values(test_results, "CPUMonitor", "powerConsumption", float),
@@ -109,13 +171,10 @@ if __name__ == '__main__':
                     ("Runtime", ""): collect_values(test_results, "NodeMonitor", "runtime", float, ns_to_s),
                 },
                 "Temperature (C)": {
+                    ("CPU", ""): collect_values(test_results, "CPUMonitor", "temperature", float),
                     ("GPU", ""): collect_values(test_results, "GPUMonitor", "temperature", float),
                 },
-                "Utilization Rate (%)": {
-                    ("CPU Core", ""): collect_values(test_results, "CPUMonitor", "coreUtilizationRate", float, to_percentage),
-                    ("GPU Core", ""): collect_values(test_results, "GPUMonitor", "coreUtilizationRate", float, to_percentage),
-                    ("GPU Memory", ""): collect_values(test_results, "GPUMonitor", "memoryUtilizationRate", float, to_percentage),
-                }
+                "Utilization Rate (%)": utilization_rate_plot
             }
 
             # Create the figure
@@ -155,7 +214,7 @@ if __name__ == '__main__':
                 for (series_name, series_style), series_values in plot_series.items():
                     axes.plot(series_values.keys(), series_values.values(), series_style, label=series_name)
 
-                    if show_final_values:
+                    if show_final_values and len(series_values.values()) > 0:
                         final_value = list(series_values.values())[-1]
                         pyplot.annotate(f"{final_value:n}", xy=(1, final_value), xytext=(8, 0), xycoords=("axes fraction", "data"), textcoords="offset points")
